♻️ Hoist manuscript format helpers to scms-core
Move manuscriptFormats and isPreviewCandidate into scms-core so both
the upload route and shared metadata utilities can import the same
predicates. Route modules become thin re-exports; tests move with them.
