+++ packages/scms-core/src/utils/index.ts
@@ -22,6 +22,7 @@ export * from './wildcard.js';
 export * from './formatZodError.js';
 export * from './versionTimelineUrls.js';
 export * from './versionTimelineTrim.js';
+export * from './manuscriptFormats.js';
 export * from './workVersionMetadata.js';
 
 export const version = 'v1';
